order items in OrderRequestDTO must be NotEmpty instead NotBlank. Remove Bean annotation from DaoAuthenticationProvider in SecurityConfig to not suffer provider duplication. convert double to BigDecimal for prices. adding in memory cache for brands and categories. implement tests. add indexes to improve search performance. Change the upload directory to an external path outside the project (e.g., app root /uploads/). This prevents loss during rebuilds. Add optimistic locking to prevent overselling(This uses a @Version). add admin email and password to properties file. Fixes for Potential NullPointer in Image Handling. Create a proper Readme file.

+++ src/main/java/com/devs/ecommerce/dto/OrderItemResponseDTO.java
@@ -1,5 +1,7 @@
 package com.devs.ecommerce.dto;
 
+import java.math.BigDecimal;
+
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Positive;
 import jakarta.validation.constraints.PositiveOrZero;
@@ -15,5 +17,5 @@ public class OrderItemResponseDTO {
     @PositiveOrZero(message = "quantity should be positive or zero")
     private Integer quantity;
     @Positive(message = "Price is always positive")
-    private double price; // Price per unit
+    private BigDecimal price; // Price per unit
 }
