+++ src/main/java/com/devs/ecommerce/dto/TownDTO.java
@@ -1,5 +1,7 @@
 package com.devs.ecommerce.dto;
 
+import java.math.BigDecimal;
+
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.PositiveOrZero;
 import lombok.Data;
@@ -10,5 +12,5 @@ public class TownDTO {
     @NotBlank(message = "Town name is required")
     private String name;
     @PositiveOrZero(message = "Delivery task must be positive or zero")
-    private double deliveryTax;
+    private BigDecimal deliveryTax;
 }
